ajout de nouvelle information

+++ ErrorApi/src/main/java/com/api/ErrorApi/Reposotory/RoleRepo.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 
-public interface RoleRepo extends JpaRepository<role, Integer> {
+public interface RoleRepo extends JpaRepository<role, Long> {
     //role findByNom(String nom);
 
 }
