+++ ErrorApi/src/main/java/com/api/ErrorApi/Reposotory/UserReposotory.java
@@ -3,7 +3,7 @@ package com.api.ErrorApi.Reposotory;
 import com.api.ErrorApi.Modele.user;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserReposotory extends JpaRepository<user, Integer> {
+public interface UserReposotory extends JpaRepository<user, Long> {
     user findByEmail(String email);
 
 }
